Allow masked to follow their target player through entrance teleports
The masked doesn't call SetMovingTowardsTargetPlayer() when it is not able to see the player. Therefore, we have to hook into NavigateTowardsTargetPlayer() to follow a smart path towards a player.

Additionally, calling SetDestinationToPosition() in GoToSmartPathDestination() was causing the masked to stop following its target player. Instead, split that method up into one which only sets the destination, and one that resets movingTowardsTargetPlayer.